refactor prices+fees
- All fees (except balancer) are specified to `1e6` (Uniswap v3 standard) as base
- Fees are now included in costs
- Removed fixed point calculations and replaced with f64 directly
- Balancer fee must be u64 (bug fix)
- Consistent error type for unsupported tokens
- Uniswap v3 had the spot pricing backwards

+++ contracts/src/Balancer/GetBalancerPoolDataBatchRequest.sol
@@ -22,7 +22,7 @@ contract GetBalancerPoolDataBatchRequest {
         uint8[] decimals;
         uint256[] liquidity;
         uint256[] weights;
-        uint32 fee;
+        uint64 fee;
     }
 
     constructor(address[] memory pools) {
@@ -62,7 +62,7 @@ contract GetBalancerPoolDataBatchRequest {
             }
 
             // Grab the swap fee
-            poolData.fee = uint32(IBPool(poolAddress).getSwapFee());
+            poolData.fee = uint64(IBPool(poolAddress).getSwapFee());
             poolData.tokens = tokens;
             poolData.decimals = decimals;
             poolData.liquidity = liquidity;
